Class Diagram Update
Added command classes to demonstrate interaction of user with program.
Certain class and their respective attributes and methods will need to
be modified in the original preliminary class diagram (could change
employee class to user class that extends Employee, Admin, etc.).

There may be some elements that are redundant in the class diagram.